Fix file modes.
https://docs.ansible.com/ansible/latest/modules/copy_module.html#parameter-mode
For those used to /usr/bin/chmod remember that modes are actually octal
numbers. You must either add a leading zero so that Ansible's YAML
parser knows it is an octal number (like 0644 or 01777)or quote it (like
'644' or '1777') so Ansible receives a string and can do its own
conversion from string into number.