Use separate return codes on failure.
Loosely based on errno(1).

The `ANSIBLE_INSTALL_METHOD` test returns the same code than a failure for
missing `--name` option, but this Ansible install pethod is a workaround
that is supposed to be removed as soon as the Ubuntu packaging for
Ansible is fixed, so we'll live with it for now.